Rename traits
Changed names if traits and relation methods, because the names too similar which could cause confusion.
